gst/vaapi: move VP9 encoder test classes to lib
