+++ test/gst-vaapi/encode/vp9.py
@@ -5,47 +5,10 @@
 ###
 
 from ....lib import *
-from ....lib.codecs import Codec
 from ....lib.gstreamer.vaapi.util import *
-from ....lib.gstreamer.vaapi.encoder import EncoderTest
+from ....lib.gstreamer.vaapi.encoder import VP9EncoderTest, VP9EncoderLPTest
 
 spec = load_test_spec("vp9", "encode", "8bit")
-
-@slash.requires(*have_gst_element("vaapivp9enc"))
-@slash.requires(*have_gst_element("vaapivp9dec"))
-class VP9EncoderBaseTest(EncoderTest):
-  def before(self):
-    super().before()
-    vars(self).update(
-      codec         = Codec.VP9,
-      gstencoder    = "vaapivp9enc",
-      gstdecoder    = "vaapivp9dec",
-      gstmediatype  = "video/x-vp9",
-      gstparser     = "vp9parse",
-      gstmuxer      = "matroskamux",
-      gstdemuxer    = "matroskademux",
-    )
-
-  def get_file_ext(self):
-    return "webm"
-
-@slash.requires(*platform.have_caps("encode", "vp9_8"))
-class VP9EncoderTest(VP9EncoderBaseTest):
-  def before(self):
-    super().before()
-    vars(self).update(
-      caps = platform.get_caps("encode", "vp9_8"),
-      lowpower  = False,
-    )
-
-@slash.requires(*platform.have_caps("vdenc", "vp9_8"))
-class VP9EncoderLPTest(VP9EncoderBaseTest):
-  def before(self):
-    super().before()
-    vars(self).update(
-      caps = platform.get_caps("vdenc", "vp9_8"),
-      lowpower  = True,
-    )
 
 class cqp(VP9EncoderTest):
   def init(self, tspec, case, ipmode, qp, quality, refmode, looplvl, loopshp):
